+++ CHANGES.rst
@@ -5,6 +5,8 @@ Improvements:
  * Integrations: Use the integrations manager provided by the homeserver admin via .well-known (#2815).
  * i18n: Add Welsh (cy).
  * SerializationService: Add deserialisation of Any.
+ * RiotSharedSettings: New class to handle user settings shared accross Riot apps.
+ * Widgets: Check user permission before opening a widget (TODO design: #2833).
 
 Changes in 0.10.2 (2019-11-15)
 ===============================================